Chucked the util functions back into reddit.py
